validações feitas

+++ src/controllers/usuariosController.js
@@ -11,27 +11,42 @@ usuariosController.listar = async (req, res) => {
     } catch (err) {
         console.log(err);
         res.status(500).redirect('/');
-    }
+    };
 };
 
 usuariosController.cadastrar = async (req, res) => {
-    const data = req.body;
+    const {usuarioID, usuarioLogin, usuarioSenha, usuarioFuncMat, usuarioSetor, usuarioStatus} = req.body;
 
-    // TODO: Validações
+    if (!{usuarioID, usuarioLogin, usuarioSenha, usuarioFuncMat, usuarioSetor, usuarioStatus}) {
+        res.status(400).json({message: 'Campos obrigatórios incompletos, preencha todos.' });
+        return;
+    } else if (await Usuarios.findByPk(usuarioID)){
+        res.status(400).json({message: 'Esse usuário já existe. Tente novamente.' });
+        return;
+    };
     
     try {
-        await Usuarios.create(data);
+        await Usuarios.create({
+            usuarioLogin: usuarioLogin,
+            usuarioSenha: usuarioSenha,
+            usuarioFuncMat: usuarioFuncMat,
+            usuarioSetor: usuarioSetor,
+            usuarioStatus: usuarioStatus
+        });
         res.status(201).redirect('/usuarios');
     } catch (err) {
         console.log(err);
         res.status(500).redirect('/');
-    }
+    };
 };
 
 usuariosController.deletar = async (req, res) => {
-    const { id } = req.params;
+    const { usuarioID, id } = req.params;
 
-    // Verificar permissão
+    if (!usuarioID) {
+        res.status(404).json({ 'erro': 'Usuário não encontrado' });
+        return;
+    };
 
     try {
         await Usuarios.destroy({where: {usuarioID: id}});
@@ -39,7 +54,7 @@ usuariosController.deletar = async (req, res) => {
     } catch (err) {
         console.log(err);
         res.status(500).redirect('/');
-    }
+    };
 };
 
 usuariosController.editar = async (req, res) => {
@@ -53,17 +68,32 @@ usuariosController.editar = async (req, res) => {
     } catch (err) {
         console.log(err);
         res.status(500).redirect('/');
-    }
+    };
 };
 
 usuariosController.atualizar = async (req, res) => {
-    const { id } = req.params;
+    const { usuarioID,id } = req.params;
     const { usuarioLogin, usuarioSenha, usuarioFuncMat, usuarioSetor, usuarioStatus } = req.body;
 
-    // TODO: Validações
+    if (!usuarioID) {
+        res.status(404).json({ 'erro': 'Usuário não encontrado!' });
+    };
+    if(usuarioLogin == usuarioID.usuarioLogin && usuarioSenha == usuarioID.usuarioSenha && usuarioFuncMat == usuarioID.usuarioFuncMat && usuarioSetor == usuarioID.usuarioSetor && usuarioStatus == usuarioID.usuarioStatus){
+        res.status(400).json({message: 'Nenhum campo atualizado. Atualize ao menos um campo para salvar as alterações.' });  
+        return;
+    };
+    if(usuarioLogin != usuarioID.usuarioLogin && await Usuarios.findAll({where : {usuarioLogin: usuarioLogin} })){
+        res.status(400).json({message: 'Este login já existe, tente novamente.' });  
+        return;
+    };
+
+    if(usuarioSenha != usuarioID.usuarioSenha && await Usuarios.findAll({where : {usuarioSenha: usuarioSenha} })){
+        res.status(400).json({message: 'Senha inválida, tente novamente.' });  
+        return;
+    };
 
     try {
-        var usuario = await Usuarios.findOne({where: {usuarioID: id}});
+        var usuario = await Usuarios.findOne({where: {usuarioID}});
         usuario.set({
             usuarioLogin: usuarioLogin || usuario.usuarioLogin,
             usuarioSenha: usuarioSenha || usuario.usuarioSenha,
